+++ core/src/config.h
@@ -70,6 +70,8 @@ public:
 
     const static QString apiUrlOverrideEnvironmentVariableName;
     const static QString apiCacheUrlOverrideEnvironmentVariableName;
+    const static QString appSecretOverrideEnvVar;
+    const static QString appVersionIdOverrideEnvVar;
 
     const static QString geolocationApiUrl;
     const static int geolocationTimeout;