+++ core/src/launcherworker.cpp
@@ -8,6 +8,7 @@
 #include <QtMath>
 #include <QSettings>
 #include <QFile>
+#include <QProcessEnvironment>
 #include <QGuiApplication>
 
 #include "cancellation/cancellationtoken.h"
@@ -334,9 +335,15 @@ void LauncherWorker::update(
         const Api& api, QNetworkAccessManager& nam,
         CancellationToken cancellationToken)
 {
+    QProcessEnvironment env;
     LocalPatcherData localData(locations);
 
     int latestAppVersion = api.getLatestAppVersion(data.patcherSecret(), cancellationToken);
+    if (env.contains(Config::appVersionIdOverrideEnvVar))
+    {
+        QString envVersionId = env.value(Config::appVersionIdOverrideEnvVar, "");
+        latestAppVersion = envVersionId.toInt();
+    }
 
     if (localData.isInstalledSpecific(latestAppVersion, data))
     {