+++ CHANGELOG.md
@@ -5,14 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
-### Changed
-- Launcher will set up a lock file before trying any api connection
-
 ### Added
+- #1354 Environment variables for overriding the app secret and version id
 - #463 Setting the display name of the Launcher at runtime
 - #1080 Checking for write errors caused by not enough disk space
 
 ### Changed
+- Launcher will set up a lock file before trying any api connection
 - Slightly better cancellation responsiveness
 
 ### Fixed